tests/utils/requirements.txt
kafka-python==2.0.2
elasticsearch==8.10.0
requests==2.28.0